requirements.txt
SpeechRecognition
pipwin
streamlit
sounddevice
gTTS
google-generativeai
python-dotenv
scipy
playsound
pydub
PyObjC

-e .